requirements.txt
gymnasium==0.29.1
stable-baselines3==2.3.2
torch>=2.0.0
numpy>=1.24.0
fastapi==0.110.0
uvicorn[standard]==0.29.0
python-socketio==5.11.1
matplotlib>=3.7.0
pygame>=2.5.0
scipy>=1.11.0
tensorboard>=2.14.0
